requirements.txt
mcp
pydantic
pillow
numpy
faiss-cpu
requests
tqdm
markitdown[all]
python-telegram-bot
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
aiohttp
python-dotenv
sse-starlette
telegram-bot
asyncio
bs4
dotenv
fastapi
httpx
llama-index
llama-index-embeddings-google-genai
mcp[cli]
pymupdf4llm
rich
telegram
trafilatura[all]